+++ cmd/clone/clone-topic.go
@@ -14,7 +14,7 @@ func newCloneTopicCmd() *cobra.Command {
 		Short: "clone existing topic (number of partitions, replication factor, config entries) to new one",
 		Args:  cobra.ExactArgs(2),
 		Run: func(cmd *cobra.Command, args []string) {
-			if !(&k8s.Operation{}).TryRun(cmd, args) {
+			if !k8s.NewOperation().TryRun(cmd, args) {
 				if err := (&topic.Operation{}).CloneTopic(args[0], args[1]); err != nil {
 					output.Fail(err)
 				}