+++ cmd/deletion/delete-acl.go
@@ -17,7 +17,7 @@ func newDeleteACLCmd() *cobra.Command {
 		Short:   "delete an acl",
 		Args:    cobra.MaximumNArgs(0),
 		Run: func(cmd *cobra.Command, args []string) {
-			if !(&k8s.Operation{}).TryRun(cmd, args) {
+			if !k8s.NewOperation().TryRun(cmd, args) {
 				if err := (&acl.Operation{}).DeleteACL(flags); err != nil {
 					output.Fail(err)
 				}